agent/requirements.txt
kafka-python-ng==2.2.3
psutil==5.9.8
pywin32==306
